adds tailor functionality
very buggy right now original clothes are not being loaded into original
and with the way i am finding crew it needs an update to the crew member
life asking them to follow to ad dit to the crew lsit but i will change
that later. but right now big issue is to get the orignal clothes back
on the crewmembers.